split components into features, rename Pdf.tsx to PDFDocument.tsx
