+++ client/src/app/invoices/page.tsx
@@ -1,4 +1,4 @@
-import InvoiceTable from '@/components/InvoiceTable';
+import InvoiceTable from '@/components/invoice/InvoiceTable';
 
 const InvoicesPage = () => {
   return <InvoiceTable />;
