+++ client/src/app/page.tsx
@@ -1,4 +1,4 @@
-import InvoiceTable from '@/components/InvoiceTable';
+import InvoiceTable from '@/components/invoice/InvoiceTable';
 
 export default function Home() {
   return <InvoiceTable />;
