+++ client/src/components/client/ClientSectionBottomContent.tsx
